Add useDataProvider Hook
